correction format paramètres memory

+++ lib/memory/memory.cpp
@@ -1,9 +1,9 @@
 #include "memory.h"
 #include <stdint.h>
 
-int comparaison(int id, int* knownlist,int taille) {  // Déclare la fonction 'comparaison' qui prend l'ID du destinataire, l'ID du paquet, un tableau d'IDs, l'ID de l'expéditeur et la taille du tableau.
+int comparaison(uint16_t id, uint16_t* knownlist,int taille) {  // Déclare la fonction 'comparaison' qui prend l'ID du destinataire, l'ID du paquet, un tableau d'IDs, l'ID de l'expéditeur et la taille du tableau.
     int known = 0;  // Déclare et initialise la variable 'known' à 0, qui indiquera si l'ID a été trouvé dans le tableau.
-    int *tmp;
+    uint8_t *tmp;
     for (int i = 0; knownlist[i] >0; i++) {  // Parcourt le tableau d'IDs jusqu'à la taille actuelle du tableau.
         if (knownlist[i] == id) {  // Si l'élément actuel du tableau correspond à l'ID cherché.
             known = 1;  // Met la variable 'tknown' à 1 si l'ID a été trouvé.
